// src/common/exceptions/custom-error-code.enum.ts
export enum CustomErrorCode {
  USER_ALREADY_EXISTS = 1001,
  USER_EMAIL_INVALID = 1002,
  USER_NAME_INVALID = 1003,
  USER_NAME_TOO_LONG = 1004,
  USER_NAME_TOO_SHORT = 1005,
  EMAIL_INVALID = 1007,
  NICKNAME_TOO_LONG = 1008,
  PHONE_TOO_LONG = 1009,
  LOGIN_USER_NOT_FOUND = 1010,
  LOGIN_USER_WRONG_PASSWORD = 1011,
  USER_PASSWORD_INVALID = 1012,
  USER_PASSWORD_TOO_SHORT = 1013,
  UNAUTHENTICATED_ERROR = 1014,
  INVALID_REFRESH_TOKEN = 1015,
  TOKEN_EXPIRED = 1016,
  NO_RESULT = 1017,
  IMAGE_NOT_FOUND = 1022,
  USER_PHONE_INVALID = 1023,
  NO_PERMISSION = 1028,
  INVALID_TABLE_CONTENT_BY_ID_ERROR = 9994,
  INVALID_EXTRACTING_TABLE_ERROR = 9995,
  INVALID_REQUEST_NULL_FIELD = 9996,
  INVALID_REQUEST_NULL = 9997,
  INVALID_DATABASE_ERROR = 9998,
  INTERNAL_SERVER_ERROR = 9999,
}